+++ src/main/java/game/gameobjects/Floor.java
@@ -29,6 +29,7 @@ import game.gamelogic.abilities.HasAbilities;
 import game.gamelogic.abilities.HasPassives;
 import game.gamelogic.abilities.Passive;
 import game.gamelogic.behavior.Behavable;
+import game.gamelogic.combat.Attack;
 import game.gamelogic.combat.AttackInfo;
 import game.gamelogic.combat.CombatModifier;
 import game.gamelogic.combat.OnAttack;
@@ -429,262 +430,7 @@ public class Floor{
     }
 
     private static AttackResult attackWithWeapon(Entity attacker, Entity defender, Weapon attackerWeapon){
-        AttackInfo attackInfo = new AttackInfo(attacker, defender, attackerWeapon);
-        int defenderDodge = 0;
-        
-        List<CombatModifier> attackerCombatMods = getCombatModifiers(attacker);
-        List<OnAttack> attackerOnAttack = new ArrayList<OnAttack>();
-        List<OnAttacked> attackerOnAttacked = new ArrayList<OnAttacked>();
-        List<OnHit> attackerOnHit = new ArrayList<OnHit>();
-        List<OnHitted> attackerOnHitted = new ArrayList<OnHitted>();
-        List<OnMiss> attackerOnMiss = new ArrayList<OnMiss>();
-        List<OnMissed> attackerOnMissed = new ArrayList<OnMissed>();
-        List<OnCrit> attackerOnCrit = new ArrayList<OnCrit>();
-        List<OnCritted> attackerOnCritted = new ArrayList<OnCritted>();
-        List<OnKill> attackerOnKill = new ArrayList<OnKill>();
-        for (CombatModifier combatModifier : attackerCombatMods) {
-            if (combatModifier instanceof OnAttack onAttack){
-                attackerOnAttack.add(onAttack);
-            }
-            if (combatModifier instanceof OnAttacked onAttacked){
-                attackerOnAttacked.add(onAttacked);
-            }
-            if (combatModifier instanceof OnHit onHit){
-                attackerOnHit.add(onHit);
-            }
-            if (combatModifier instanceof OnHitted onHitted){
-                attackerOnHitted.add(onHitted);
-            }
-            if (combatModifier instanceof OnMiss onMiss){
-                attackerOnMiss.add(onMiss);
-            }
-            if (combatModifier instanceof OnMissed onMissed){
-                attackerOnMissed.add(onMissed);
-            }
-            if (combatModifier instanceof OnCrit onCrit){
-                attackerOnCrit.add(onCrit);
-            }
-            if (combatModifier instanceof OnCritted onCritted){
-                attackerOnCritted.add(onCritted);
-            }
-            if (combatModifier instanceof OnKill onKill) {
-                attackerOnKill.add(onKill);
-            }
-        }
-
-        List<CombatModifier> defenderCombatMods = getCombatModifiers(defender);
-        List<OnAttack> defenderOnAttack = new ArrayList<OnAttack>();
-        List<OnAttacked> defenderOnAttacked = new ArrayList<OnAttacked>();
-        List<OnHit> defenderOnHit = new ArrayList<OnHit>();
-        List<OnHitted> defenderOnHitted = new ArrayList<OnHitted>();
-        List<OnMiss> defenderOnMiss = new ArrayList<OnMiss>();
-        List<OnMissed> defenderOnMissed = new ArrayList<OnMissed>();
-        List<OnCrit> defenderOnCrit = new ArrayList<OnCrit>();
-        List<OnCritted> defenderOnCritted = new ArrayList<OnCritted>();
-        List<OnDeath> defenderOnDeath = new ArrayList<OnDeath>();
-        for (CombatModifier combatModifier : defenderCombatMods) {
-            if (combatModifier instanceof OnAttack onAttack){
-                defenderOnAttack.add(onAttack);
-            }
-            if (combatModifier instanceof OnAttacked onAttacked){
-                defenderOnAttacked.add(onAttacked);
-            }
-            if (combatModifier instanceof OnHit onHit){
-                defenderOnHit.add(onHit);
-            }
-            if (combatModifier instanceof OnHitted onHitted){
-                defenderOnHitted.add(onHitted);
-            }
-            if (combatModifier instanceof OnMiss onMiss){
-                defenderOnMiss.add(onMiss);
-            }
-            if (combatModifier instanceof OnMissed onMissed){
-                defenderOnMissed.add(onMissed);
-            }
-            if (combatModifier instanceof OnCrit onCrit){
-                defenderOnCrit.add(onCrit);
-            }
-            if (combatModifier instanceof OnCritted onCritted){
-                defenderOnCritted.add(onCritted);
-            }
-            if (combatModifier instanceof OnDeath onDeath){
-                defenderOnDeath.add(onDeath);
-            }
-        }
-
-        defenderDodge += getDodge(defender);
-
-        attackInfo.setDefenderDodge(defenderDodge);
-
-        int naturalAttackerRoll = App.randomNumber(1, 20);
-
-        attackInfo.setBaseRoll(naturalAttackerRoll);
-
-        boolean crit = naturalAttackerRoll == 20;
-
-        attackInfo.setCrit(crit);
-
-        int modifiedAttackerRoll = naturalAttackerRoll;
-
-        modifiedAttackerRoll += getAccuracy(attacker, attackerWeapon);
-
-        attackInfo.setModifiedRoll(modifiedAttackerRoll);
-        
-        int damage = attackerWeapon.generateDamage();
-
-        damage += Attribute.getAttribute(Attribute.STRENGTH, attacker);
-        
-        damage = crit ? damage * 2 : damage;
-
-        attackInfo.setDamage(damage);
-
-        DamageType attackerDamageType = attackerWeapon.getDamageType();
-        
-        attackInfo.setDamageType(attackerDamageType);
-
-        boolean hit = modifiedAttackerRoll >= defenderDodge;
-
-        attackInfo.setHit(hit);
-        
-        int damageDelt = 0;
-
-        attackInfo.setDamageDelt(damageDelt);
-
-        // attacker on attack and defender on attacked
-        for (OnAttack onAttack : attackerOnAttack) {
-            onAttack.doOnAttack(attacker, defender, attackInfo);
-        }
-        for (OnAttacked onAttacked : defenderOnAttacked) {
-            onAttacked.doOnAttacked(defender, attacker, attackInfo);
-        }
-        //
-        
-        if (hit){
-
-            if (crit) {
-                if (attacker instanceof PlayerEntity){
-                    Display.log("Critical hit!");
-                } else if (defender instanceof PlayerEntity){
-                    Display.log("The " + attacker.getName() + " scores a critical hit on you.");
-                } else {
-                    Display.log("The " + attacker.getName() + " scores a critical hit on the " + defender.getName() + ".", attacker.getSpace());
-                }
-            }
-
-            damageDelt = defender.dealDamage(damage, attackerDamageType, attacker);
-
-            attackInfo.setDamageDelt(damageDelt);
-
-            // attacker on hit and defender on hitted
-            for (OnHit onHit : attackerOnHit) {
-                onHit.doOnHit(attacker, defender, attackInfo);
-            }
-            for (OnHitted onHitted : defenderOnHitted) {
-                onHitted.doOnHitted(defender, attacker, attackInfo);
-            }
-            //
-            
-            if (crit){
-
-                // attacker on crit and defender on critted
-                for (OnCrit onCrit : attackerOnCrit) {
-                    onCrit.doOnCrit(attacker, defender, attackInfo);
-                }
-                for (OnCritted onCritted : defenderOnCritted) {
-                    onCritted.doOnCritted(defender, attacker, attackInfo);
-                }
-
-            }
-
-            if (!defender.isAlive()) {
-                for (OnDeath onDeath : defenderOnDeath) {
-                    onDeath.doOnDeath(defender, attacker, attackInfo);
-                }
-                for (OnKill onKill : attackerOnKill) {
-                    onKill.doOnKill(attacker, defender, attackInfo);
-                }
-            }
-
-        } else {
-            if (attacker instanceof PlayerEntity){
-                Display.log("You miss the " + defender.getName() + ".");
-            } else if (defender instanceof PlayerEntity){
-                Display.log("The " + attacker.getName() + " misses you.");
-            } else {
-                Display.log(attacker.getName() + " misses the " + defender.getName() + ".", defender.getSpace());
-            }
-
-            // attacker on miss and defender on missed
-            for (OnMiss onMiss : attackerOnMiss) {
-                onMiss.doOnMiss(attacker, defender, attackInfo);
-            }
-            for (OnMissed onMissed : defenderOnMissed) {
-                onMissed.doOnMissed(defender, attacker, attackInfo);
-            }
-            //
-        }
-        return new AttackResult(hit, crit, damageDelt, attackerDamageType, attacker, defender);
-    }
-
-    
-    public static List<CombatModifier> getCombatModifiers(Entity entity){
-        return App.recursiveCheck(entity, getCombatModConditions(), (obj) -> {
-            if (obj instanceof CombatModifier cm){
-                return Optional.of(cm);
-            }
-            return Optional.empty();
-        });
-    }
-
-    private static CheckConditions getCombatModConditions(){
-        return CheckConditions.all()
-            .withInventory(false);
-    }
-
-    public static int getDodge(Entity entity){
-        int dodge = 0;
-        for (ModifiesDodge modifiesDodge : App.recursiveCheck(entity, getDodgeConditions(), (obj) -> {
-            if (obj instanceof ModifiesDodge md) {
-                return Optional.of(md);
-            }
-            return Optional.empty();
-        })) {
-            dodge = modifiesDodge.modifyDodge(dodge);
-        }
-        return dodge;
-    }
-
-    private static CheckConditions getDodgeConditions(){
-        return CheckConditions.all()
-            .withInventory(false);
-    }
-    
-    public static int getAccuracy(Entity entity, Weapon activeWeapon){
-
-        int accuracy = 0;
-
-        for (ModifiesAccuracy modifiesAccuracy : App.recursiveCheck(entity, getAccuracyConditions(), (obj) -> {
-            if (obj instanceof ModifiesAccuracy ma) {
-                return Optional.of(ma);
-            }
-            return Optional.empty();
-        })){
-            accuracy = modifiesAccuracy.modifyAccuracy(accuracy);
-        };
-
-        if (activeWeapon instanceof ModifiesAccuracy ma) {
-            accuracy = ma.modifyAccuracy(accuracy);
-        }
-
-        return accuracy;
-
-    }
-
-    private static CheckConditions getAccuracyConditions(){
-        return CheckConditions.all()
-            .withInventory(false)
-            .withArmedWeapons(false)
-            .withUnarmedWeapon(false);
+        return new Attack(attacker, defender, attackerWeapon).execute();
     }
 
 }
